requirements.txt
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
dash==2.0.0
dash-bootstrap-components==1.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.2.0
dash-table==5.0.0
Flask==1.1.1
Flask-Compress==1.4.0
fonttools==4.37.4
future==0.18.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==2.10.3
kaleido==0.2.1
kiwisolver==1.4.4
MarkupSafe==1.1.1
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
plotly==5.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2019.3
retrying==1.3.3
six==1.13.0
tenacity==8.1.0
Werkzeug==0.16.0
scikit-learn
openpyxl
seaborn
keras-models